requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and RAG
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-text-splitters>=0.1.0

# OpenAI
openai>=1.10.0

# Vector store
chromadb>=0.4.22

# Text processing
beautifulsoup4>=4.12.0
pandas>=2.0.0
tiktoken>=0.5.0

# Utilities
lxml>=4.9.0

# UI Components
streamlit-extras>=0.7.8
streamlit-shadcn-ui>=0.1.19
